scripts/poker_mtt/requirements.txt
PyYAML>=6.0
requests>=2.32
redis>=5.0
websocket-client>=1.8
